Refactor useRolesState
