+++ src/hooks/DAO/loaders/useHatsTree.ts
@@ -8,7 +8,7 @@ import { StreamsQueryDocument } from '../../../../.graphclient';
 import { SablierPayment } from '../../../components/pages/Roles/types';
 import useIPFSClient from '../../../providers/App/hooks/useIPFSClient';
 import { useNetworkConfig } from '../../../providers/NetworkConfig/NetworkConfigProvider';
-import { DecentHatsError, useRolesState } from '../../../state/useRolesState';
+import { useRolesState, DecentHatsError } from '../../../state/roles';
 import { CacheExpiry, CacheKeys } from '../../utils/cache/cacheDefaults';
 import { getValue, setValue } from '../../utils/cache/useLocalStorage';
 
@@ -17,7 +17,9 @@ const hatsSubgraphClient = new HatsSubgraphClient({
 });
 
 const useHatsTree = () => {
-  const { hatsTreeId, hatsTree, streamsFetched, setHatsTree, updateRolesWithStreams } = useRolesState();
+  const { hatsTreeId, hatsTree, streamsFetched, setHatsTree, updateRolesWithStreams } =
+    useRolesState();
+
   const ipfsClient = useIPFSClient();
   const {
     chain,
