+++ src/pages/daos/[daoAddress]/roles/edit/index.tsx
@@ -22,7 +22,7 @@ import { useRolesSchema } from '../../../../../hooks/schemas/roles/useRolesSchem
 import useCreateRoles from '../../../../../hooks/utils/useCreateRoles';
 import { useFractal } from '../../../../../providers/App/AppProvider';
 import { useNetworkConfig } from '../../../../../providers/NetworkConfig/NetworkConfigProvider';
-import { useRolesState } from '../../../../../state/useRolesState';
+import { useRolesState } from '../../../../../state/roles';
 
 function RolesEdit() {
   const { t } = useTranslation(['roles', 'navigation', 'modals', 'common']);
